+++ lib/Listeners/DocumentUnsavedListener.php
@@ -1,0 +1,94 @@
+<?php
+
+/**
+ *
+ * (c) Copyright Ascensio System SIA 2025
+ *
+ * This program is a free software product.
+ * You can redistribute it and/or modify it under the terms of the GNU Affero General Public License
+ * (AGPL) version 3 as published by the Free Software Foundation.
+ * In accordance with Section 7(a) of the GNU AGPL its Section 15 shall be amended to the effect
+ * that Ascensio System SIA expressly excludes the warranty of non-infringement of any third-party rights.
+ *
+ * This program is distributed WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * For details, see the GNU AGPL at: http://www.gnu.org/licenses/agpl-3.0.html
+ *
+ * You can contact Ascensio System SIA at 20A-12 Ernesta Birznieka-Upisha street, Riga, Latvia, EU, LV-1050.
+ *
+ * The interactive user interfaces in modified source and object code versions of the Program
+ * must display Appropriate Legal Notices, as required under Section 5 of the GNU AGPL version 3.
+ *
+ * Pursuant to Section 7(b) of the License you must retain the original Product logo when distributing the program.
+ * Pursuant to Section 7(e) we decline to grant you any rights under trademark law for use of our trademarks.
+ *
+ * All the Product's GUI elements, including illustrations and icon sets, as well as technical
+ * writing content are licensed under the terms of the Creative Commons Attribution-ShareAlike 4.0 International.
+ * See the License terms at http://creativecommons.org/licenses/by-sa/4.0/legalcode
+ *
+ */
+
+namespace OCA\Onlyoffice\Listeners;
+
+use OCA\Onlyoffice\AppInfo\Application;
+use OCA\Onlyoffice\Events\DocumentUnsavedEvent;
+use OCP\EventDispatcher\Event;
+use OCP\EventDispatcher\IEventListener;
+use OCP\IConfig;
+use OCP\L10N\IFactory;
+use OCP\Notification\IManager as NotificationIManager;
+use OCP\Notification\IncompleteNotificationException;
+use OCP\Notification\InvalidValueException;
+use Psr\Log\LoggerInterface;
+
+/**
+ * Document Unsaved Event Listener
+ */
+class DocumentUnsavedListener implements IEventListener {
+
+    /**
+     * Translation object
+     * @var
+     */
+    private $trans;
+    
+    /**
+     * @param \OCP\IConfig $config
+     * @param \Psr\Log\LoggerInterface $logger
+     * @param \OCP\Notification\IManager $notificationManager
+     * @param \OCP\L10N\IFactory $l10nFactory
+     */
+    public function __construct(
+        private IConfig $config,
+        private LoggerInterface $logger,
+        private NotificationIManager $notificationManager,
+        private IFactory $l10nFactory,
+    ) {}
+
+    public function handle(Event $event): void {
+        if (!$event instanceof DocumentUnsavedEvent) {
+            return;
+        }
+
+        $lang = $this->config->getUserValue($event->getUserId(), 'core', 'lang', 'en');
+        $this->trans = $this->l10nFactory->get(Application::class, $lang);
+        $this->notifySender($event->getUserId(), $event->getFileId(), $event->getFileName());
+    }
+
+    private function notifySender($uid, $fileId, $fileName) {
+        try {
+            $notification = $this->notificationManager->createNotification();
+            $notification->setApp(Application::APP_ID)
+                ->setDateTime(new \DateTime())
+                ->setObject("document_unsaved", $fileId)
+                ->setSubject("document_unsaved", [
+                    'fileId' => $fileId,
+                    'fileName' => $fileName,
+                ])
+                ->setUser($uid);
+            $this->notificationManager->notify($notification);
+        } catch (InvalidValueException | IncompleteNotificationException $e) {
+            $this->logger->error("Invalid value for notification provided: " . $e->getMessage(), ['exception' => $e]);
+        }
+    }
+}