+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ## Added
 - plugin description and useful links in admin settings
+- document unsaved notifications
 
 ## 9.11.0
 ## Changed